Decouple restoring Playable from Preferences and storing position
Breaks dependency cycle
